requirements.txt
acme==0.38.0
asn1crypto==0.24.0
aspy.yaml==1.3.0
astroid==2.2.5
certbot==0.38.0
certifi==2019.9.11
cffi==1.12.3
cfgv==2.0.1
chardet==3.0.4
ConfigArgParse==0.14.0
configobj==5.0.6
cryptography==2.7
distro==1.4.0
Django==2.2.1
future==0.17.1
Glances==3.1.0
identify==1.4.6
idna==2.8
importlib-metadata==0.19
isort==4.3.20
josepy==1.2.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
nodeenv==1.3.3
parsedatetime==2.4
pre-commit==1.18.2
psutil==5.6.2
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
pyRFC3339==1.1
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.3.5
urllib3==1.25.6
virtualenv==16.7.4
wrapt==1.11.1
zipp==0.6.0
zope.component==4.5
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.proxy==4.3.2
